refactor(widgetbook): split diplomacy detail catalog part (Refs #3753)
Extract GAME30002 Widgetbook stories into catalog_diplomacy_detail.dart
so catalog_data_screens.dart stays under the repo.part_unit_size limit.
