+++ app/lib/widgetbook/catalog.dart
@@ -129,6 +129,7 @@ import '../widgets/ct_transfer_list.dart';
 part 'catalog_panel_map_stories.dart';
 part 'catalog_panels.dart';
 part 'catalog_diplomacy_panel.dart';
+part 'catalog_diplomacy_detail.dart';
 part 'catalog_screens_combat.dart';
 part 'catalog_dialogs.dart';
 part 'catalog_primitives.dart';
